LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.4_Comments/S7.4_A4_T4-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Unexpected token '/'
S7.4_A4_T4

PASS Expected parsing failure

TEST COMPLETE
